Add tests and refactor.

+++ beacon_node/beacon_chain/src/test_utils.rs
@@ -777,7 +777,7 @@ where
         self.chain
             .data_availability_checker
             .custody_context()
-            .num_of_data_columns_to_sample(None, &self.chain.spec) as usize
+            .num_of_data_columns_to_sample(None, &self.chain.spec)
     }
 
     pub fn slots_per_epoch(&self) -> u64 {
